FST: Optimise and simplify the RAIN read implementation by using the helper functions      to compute local and original positions. This makes the normal read similar to      the future implementation of vector reads.      Use "uint64_t" for offset values and "uint32_t" for length values throughout the      RAIN layouts      Reformat some of the debug messages and drop all the code that is no longer used

+++ fst/layout/RaidDpLayout.hh
@@ -119,7 +119,7 @@ private:
   //! @param length data length
   //!
   //--------------------------------------------------------------------------
-  virtual void AddDataBlock (off_t offset, const char* buffer, size_t length);
+  virtual void AddDataBlock (uint64_t offset, const char* buffer, uint32_t length);
 
 
   //--------------------------------------------------------------------------
@@ -139,7 +139,7 @@ private:
   //! @return 0 if successful, otherwise error
   //!
   //--------------------------------------------------------------------------
-  virtual int WriteParityToFiles (off_t offsetGroup);
+  virtual int WriteParityToFiles (uint64_t offsetGroup);
 
 
   //--------------------------------------------------------------------------
@@ -167,9 +167,9 @@ private:
   //! @return true if successful, otherwise error
   //!
   //--------------------------------------------------------------------------
-  bool RecoverPiecesInGroup (off_t offsetInit,
+  bool RecoverPiecesInGroup (uint64_t offsetInit,
                              char* pBuffer,
-                             std::map<off_t, size_t>& rMapPieces);
+                             std::map<uint64_t, uint32_t>& rMapPieces);
 
 
   //--------------------------------------------------------------------------
@@ -277,14 +277,14 @@ private:
 
   //--------------------------------------------------------------------------
   //! Convert a global offset (from the inital file) to a local offset within
-  //! a stripe file. The initial block does *NOT* span multiple chunks (stripes)
-  //! therefore if the original length is bigger than one chunk the splitting
-  //! must be done before calling this method.
+  //! a stripe data file. The initial block does *NOT* span multiple chunks
+  //! (stripes) therefore if the original length is bigger than one chunk the
+  //! splitting must be done before calling this method.
   //!
   //! @param global_off initial offset
   //!
-  //! @return tuple made up of the logical index of the stripe file the piece
-  //!         belongs to and the local offset within that file. 
+  //! @return tuple made up of the logical index of the stripe data file the
+  //!         piece belongs to and the local offset within that file. 
   //!
   //--------------------------------------------------------------------------
   virtual std::pair<int, uint64_t>
@@ -292,8 +292,10 @@ private:
 
 
   //--------------------------------------------------------------------------
-  //! Convert a local position (from a stripe file) to a global position
-  //! within the initial file file
+  //! Convert a local position (from a stripe data file) to a global position
+  //! within the initial file file. Note that the local offset has to come
+  //! from a stripe data file since there is no corresponde in the original
+  //! file for a piece which is in the parity stripe.
   //!
   //! @param stripe_id logical stripe index
   //! @param local_off local offset
